♻️ refactor: moved useUser into page.tsx
